remove agents
